# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Lavoro Ghana Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (admin)
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}/suspend:
    patch:
      operationId: suspendUser
      tags: [users]
      summary: Suspend or unsuspend a user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuspendInput"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # --- FREELANCER PROFILES ---
...
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review for a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/user/{userId}:
    get:
      operationId: listUserReviews
      tags: [reviews]
      summary: Get reviews for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

  # --- REPORTS ---
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List reports (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      summary: Submit a report about a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Report submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}/resolve:
    patch:
      operationId: resolveReport
      tags: [reports]
      summary: Resolve a report (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportResolveInput"
      responses:
        "200":
          description: Report resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

  # --- STATS (Dashboard data) ---
...
  /stats/top-freelancers:
    get:
      operationId: getTopFreelancers
      tags: [stats]
      summary: Get top-rated freelancers for homepage
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top freelancers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FreelancerProfile"
  /stats/featured-jobs:
    get:
      operationId: getFeaturedJobs
      tags: [stats]
      summary: Get recent/featured job listings for homepage
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Featured jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
  /stats/featured-products:
    get:
      operationId: getFeaturedProducts
      tags: [stats]
      summary: Get recent/featured products for homepage
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      required: [id, name, email, phone, role, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [freelancer, client, seller, admin]
        status:
          type: string
          enum: [active, suspended]
        avatarUrl:
          type: ["string", "null"]
        isVerified:
          type: boolean
        averageRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
    UserListResponse:
      type: object
      required: [users, total]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
    SuspendInput:
      type: object
      required: [suspended]
      properties:
        suspended:
          type: boolean
        reason:
          type: ["string", "null"]

    # --- FREELANCER PROFILES ---
    FreelancerProfile:
      type: object
      required: [id, userId, user, skills, location, hourlyRate, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        bio:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        location:
          type: string
        portfolioItems:
          type: array
          items:
            $ref: "#/components/schemas/PortfolioItem"
        experience:
          type: ["string", "null"]
        hourlyRate:
          type: ["number", "null"]
        projectRate:
          type: ["number", "null"]
        completedJobs:
          type: integer
        verificationStatus:
          type: string
          enum: [unverified, pending, verified]
        trustLabel:
          type: ["string", "null"]
        createdAt:
          type: string
...
    PortfolioItem:
      type: object
      required: [id, title]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        link:
          type: ["string", "null"]
...
    Job:
      type: object
      required: [id, clientId, client, title, description, category, budget, location, status, deadline, applicationCount, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        client:
          $ref: "#/components/schemas/User"
        title:
          type: string
        description:
          type: string
        category:
          type: string
        budget:
          type: number
        location:
          type: string
        status:
          type: string
          enum: [open, in_progress, completed, cancelled]
        deadline:
          type: string
        applicationCount:
          type: integer
        createdAt:
          type: string
...
    Product:
      type: object
      required: [id, sellerId, seller, name, description, category, price, location, quantity, negotiable, status, createdAt]
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        seller:
          $ref: "#/components/schemas/User"
        name:
          type: string
        description:
          type: string
        category:
          type: string
        price:
          type: number
        location:
          type: string
        quantity:
          type: integer
        negotiable:
          type: boolean
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [available, sold_out, removed]
        createdAt:
          type: string
...
    Review:
      type: object
      required: [id, reviewerId, revieweeId, reviewer, rating, createdAt]
      properties:
        id:
          type: integer
        reviewerId:
          type: integer
        revieweeId:
          type: integer
        reviewer:
          $ref: "#/components/schemas/User"
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [revieweeId, rating]
      properties:
        revieweeId:
          type: integer
        rating:
          type: integer
        comment:
          type: string

    # --- REPORTS ---
    Report:
      type: object
      required: [id, reporterId, reportedUserId, reason, status, createdAt]
      properties:
        id:
          type: integer
        reporterId:
          type: integer
        reportedUserId:
          type: integer
        reason:
          type: string
        details:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, resolved, dismissed]
        resolution:
          type: ["string", "null"]
        createdAt:
          type: string
    ReportInput:
      type: object
      required: [reportedUserId, reason]
      properties:
        reportedUserId:
          type: integer
        reason:
          type: string
        details:
          type: string
    ReportResolveInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [resolved, dismissed]
        resolution:
          type: string

    # --- STATS ---
...